:sparkles: Feat: 추천 관련 로직 api 초안 완료

+++ src/main/java/yegam/opale_be/domain/chat/room/dto/response/ChatRoomUpdateDto.java
@@ -32,5 +32,5 @@ public class ChatRoomUpdateDto {
   private Boolean isActive;
 
   @Schema(description = "누적 방문자 수", example = "128")
-  private Integer visitCount;
+  private Long visitCount;
 }
